Move out login, base methods, and helper functions to there own classes
